(bugfix) Arrumando erro na tela de chat e themes
