+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
 
     <application
         android:allowBackup="true"
@@ -11,26 +11,32 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.PurpuraApp"
-        tools:targetApi="31">
+        tools:targetApi="31" >
         <activity
-            android:name=".MyProducts"
+            android:name=".ui.screens.errors.GenericError"
+            android:exported="false" />
+        <activity
+            android:name=".ui.screens.IndividualChat"
+            android:exported="false" />
+        <activity
+            android:name=".ui.screens.accountFeatures.MyProducts"
             android:exported="false"
             android:theme="@style/Theme.PurpuraApp" />
         <activity
-            android:name=".ui.screens.MyOrders"
+            android:name=".ui.screens.accountFeatures.MyOrders"
             android:exported="false"
             android:theme="@style/Theme.PurpuraApp" />
         <activity
-            android:name=".ui.screens.AddicionalInformacionsRegisterGoogle"
+            android:name=".ui.screens.autentication.AddicionalInformacionsRegisterGoogle"
             android:exported="false" />
         <activity
-            android:name=".ui.screens.RegisterProduct"
+            android:name=".ui.screens.productRegister.RegisterProduct"
             android:exported="false" />
         <activity
-            android:name=".ui.screens.EditPixKeys"
+            android:name=".ui.screens.accountFeatures.EditPixKeys"
             android:exported="false" />
         <activity
-            android:name=".ui.screens.EditAdresses"
+            android:name=".ui.screens.accountFeatures.EditAdresses"
             android:exported="false" />
         <activity
             android:name=".ui.screens.ProductPage"
@@ -45,25 +51,25 @@
             android:name=".ui.screens.ChoosePaymentMethod"
             android:exported="false" />
         <activity
-            android:name=".ui.screens.Login"
+            android:name=".ui.screens.autentication.Login"
             android:exported="false" />
         <activity
-            android:name=".ui.screens.RegisterAdress"
+            android:name=".ui.screens.productRegister.RegisterAdress"
             android:exported="false" />
         <activity
-            android:name=".ui.screens.RegisterProductEndPage"
+            android:name=".ui.screens.productRegister.RegisterProductEndPage"
             android:exported="false" />
         <activity
-            android:name=".ui.screens.RegisterPixKey"
+            android:name=".ui.screens.productRegister.RegisterPixKey"
             android:exported="false" />
         <activity
-            android:name=".ui.screens.RegisterOrLogin"
+            android:name=".ui.screens.autentication.RegisterOrLogin"
             android:exported="false" />
         <activity
-            android:name=".ui.screens.InternetError"
+            android:name=".ui.screens.errors.InternetError"
             android:exported="false" />
         <activity
-            android:name=".ui.screens.Register"
+            android:name=".ui.screens.autentication.Register"
             android:exported="false" />
         <activity
             android:name=".ui.screens.MainActivity"
@@ -72,7 +78,7 @@
             android:name=".ui.screens.SplashScreen"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/semBarra">
+            android:theme="@style/semBarra" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
