+++ app/src/main/java/com/purpura/app/database/firebase/FirebaseMethods.java
@@ -1,8 +1,7 @@
 package com.purpura.app.database.firebase;
 
 import com.google.firebase.auth.FirebaseAuth;
-import com.purpura.app.ui.screens.RegisterOrLogin;
-import com.purpura.app.configuration.Methods;
+
 public class FirebaseMethods {
 
     public void logout() {
