+++ app/src/main/java/com/purpura/app/ui/screens/IndividualChat.java
@@ -1,7 +1,8 @@
 package com.purpura.app.ui.screens;
 
 import android.os.Bundle;
-import android.widget.ImageView;
+import android.webkit.WebView;
+import android.webkit.WebViewClient;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
@@ -10,27 +11,22 @@ import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
 import com.purpura.app.R;
-import com.purpura.app.configuration.Methods;
-import com.purpura.app.ui.account.AccountFragment;
 
-public class EditAdresses extends AppCompatActivity {
-
-    Methods methods = new Methods();
+public class IndividualChat extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_edit_adresses);
+        setContentView(R.layout.activity_individual_chat);
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
             return insets;
         });
 
-        ImageView backButton = findViewById(R.id.editAdressesBackButton);
-
-        backButton.setOnClickListener(v -> methods.openScreenActivity(this, AccountFragment.class));
-
+        WebView webView = findViewById(R.id.chatWebView);
+        webView.setWebViewClient(new WebViewClient());
+        webView.loadUrl("https://purpura-react-site.onrender.com/");
     }
 }
