+++ app/src/main/java/com/purpura/app/ui/screens/productRegister/RegisterPixKey.java
@@ -1,6 +1,8 @@
-package com.purpura.app.ui.screens;
+package com.purpura.app.ui.screens.productRegister;
 
 import android.os.Bundle;
+import android.widget.Button;
+import android.widget.ImageView;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
@@ -8,19 +10,30 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
+import com.purpura.app.configuration.Methods;
 import com.purpura.app.R;
+import com.purpura.app.ui.account.AccountFragment;
 
-public class RegisterProductEndPage extends AppCompatActivity {
+public class RegisterPixKey extends AppCompatActivity {
+
+    Methods methods = new Methods();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_register_product_end_page);
+        setContentView(R.layout.activity_register_pix_key);
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
             return insets;
         });
+
+        ImageView backButton = findViewById(R.id.registerAdressBackButton);
+        Button continueButton = findViewById(R.id.registerPixKeyAddPixKeyButton);
+
+        backButton.setOnClickListener(v -> methods.openScreenActivity(this, RegisterPixKey.class));
+        continueButton.setOnClickListener(v -> methods.openScreenActivity(this, AccountFragment.class));
+
     }
 }
