+++ app/src/main/java/com/purpura/app/ui/screens/productRegister/RegisterProduct.java
@@ -1,0 +1,38 @@
+package com.purpura.app.ui.screens.productRegister;
+
+import android.os.Bundle;
+import android.widget.Button;
+import android.widget.ImageView;
+
+import androidx.activity.EdgeToEdge;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.graphics.Insets;
+import androidx.core.view.ViewCompat;
+import androidx.core.view.WindowInsetsCompat;
+
+import com.purpura.app.R;
+import com.purpura.app.configuration.Methods;
+import com.purpura.app.ui.account.AccountFragment;
+
+public class RegisterProduct extends AppCompatActivity {
+
+    Methods methods = new Methods();
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        EdgeToEdge.enable(this);
+        setContentView(R.layout.activity_register_product);
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.constraintLayout), (v, insets) -> {
+            Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
+            return insets;
+        });
+
+        ImageView backButton = findViewById(R.id.registerProductBackButton);
+        Button continueButton = findViewById(R.id.registerProductAddProductButton);
+
+        backButton.setOnClickListener(v -> methods.openScreenActivity(this, AccountFragment.class));
+        continueButton.setOnClickListener(v -> methods.openScreenActivity(this, RegisterAdress.class));
+    }
+}
