+++ app/src/main/res/layout-land/activity_login.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screens.Register">
+    tools:context=".ui.screens.autentication.Register">
 
     <com.google.android.gms.common.SignInButton
         android:id="@+id/loginWithGoogle"
