+++ app/src/main/res/layout/activity_addicional_informacions_register_google.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screens.Register"
+    tools:context=".ui.screens.autentication.Register"
     >
 
     <TextView
