+++ app/src/main/res/layout/activity_edit_adresses.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screens.EditAdresses">
+    tools:context=".ui.screens.accountFeatures.EditAdresses">
 
     <ImageView
         android:id="@+id/editAdressesBackButton"
