+++ app/src/main/res/layout/activity_edit_pix_keys.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screens.EditPixKeys">
+    tools:context=".ui.screens.accountFeatures.EditPixKeys">
 
     <ImageView
         android:id="@+id/editPixKeyBackButton"
