+++ app/src/main/res/layout/activity_internet_error.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screens.InternetError">
+    tools:context=".ui.screens.errors.InternetError">
 
     <TextView
         android:id="@+id/InternetErrorText"
