+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screens.Login">
+    tools:context=".ui.screens.autentication.Login">
 
     <com.google.android.gms.common.SignInButton
         android:id="@+id/loginWithGoogle"
