+++ app/src/main/res/layout/activity_my_orders.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screens.MyOrders">
+    tools:context=".ui.screens.accountFeatures.MyOrders">
 
     <ImageView
         android:id="@+id/myOrdersBackButton"
