+++ app/src/main/res/layout/activity_my_products.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".MyProducts">
+    tools:context=".ui.screens.accountFeatures.MyProducts">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
