+++ app/src/main/res/layout/activity_register_adress.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screens.RegisterAdress">
+    tools:context=".ui.screens.productRegister.RegisterAdress">
 
     <ImageView
         android:id="@+id/registerAdressBackButton"
