+++ app/src/main/res/layout/activity_register_or_login.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screens.RegisterOrLogin">
+    tools:context=".ui.screens.autentication.RegisterOrLogin">
 
     <TextView
         android:id="@+id/textView2"
