+++ app/src/main/res/layout/activity_register_pix_key.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screens.RegisterPixKey">
+    tools:context=".ui.screens.productRegister.RegisterPixKey">
 
     <ImageView
         android:id="@+id/registerAdressBackButton"
