+++ app/src/main/res/layout/activity_register_product.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:theme="@style/semBarra"
-    tools:context=".ui.screens.RegisterProduct">
+    tools:context=".ui.screens.productRegister.RegisterProduct">
 
     <ScrollView
         android:layout_width="match_parent"
