+++ app/src/main/res/layout/activity_register_product_end_page.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screens.RegisterProductEndPage">
+    tools:context=".ui.screens.productRegister.RegisterProductEndPage">
 
     <ImageView
         android:id="@+id/registerAdressBackButton"
