+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -16,7 +16,7 @@
         android:title="Carrinho" />
 
     <item
-        android:id="@+id/chatFragment"
+        android:id="@+id/chatListFragment"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:icon="@drawable/chat_icon"
